backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
langchain==0.3.0
langchain-openai==0.2.2  # Fixed compatible version
celery==5.4.0
redis==5.0.7
psycopg2-binary==2.9.9
pandas==2.2.2
transformers==4.42.3
torch==2.3.1
sendgrid==6.11.0
python-dotenv==1.0.1